requirements.txt
matplotlib
numpy
opencv-python
pdbpp
scipy
Pillow
gdown
protobuf>=3.20.3,<5.0.0dev
chardet
types-requests
pytype
dynaconf
